# packages/api/src/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "ayni API",
    "version": "0.4.0",
    "description": "Endpoints HTTP que usa el SDK de Ayni para sincronizar workflows y modelos.\n\n## Autenticación\n\nCada solicitud lleva el secreto de una credencial del SDK en el encabezado:\n\n```http\nAuthorization: Bearer ayni_sk_…\n```\n\n- Un administrador genera la credencial en el dashboard, en Credenciales SDK de la aplicación. El secreto se muestra una sola vez.\n- La credencial solo da acceso a los recursos de su aplicación.\n- Cuando un administrador revoca o regenera la credencial, estos endpoints responden `401` con `credentialRevoked`. Revocarla no borra workflows, modelos ni versiones, ni lo que ya está instalado en los dispositivos."
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Local development server"
    }
  ],
  "components": {
    "securitySchemes": {
      "sdkCredential": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "ayni_sk_…",
        "description": "Secreto de una credencial del SDK: una cadena opaca que empieza por `ayni_sk_`, no un JWT. El servidor no lee nada dentro del secreto; lo compara con el hash de las credenciales activas de la aplicación."
      },
      "datasetUserSession": {
        "type": "apiKey",
        "in": "cookie",
        "name": "better-auth.session_token",
        "description": "Cookie de sesión de Better Auth para las rutas del dashboard."
      },
      "userSession": {
        "type": "apiKey",
        "in": "cookie",
        "name": "better-auth.session_token",
        "description": "Cookie de sesión de Better Auth para las rutas del dashboard."
      }
    },
    "schemas": {
...
      "SdkTraceArtifactUploadIntent": {
        "type": "object",
        "properties": {
          "artifactId": {
            "type": "string",
            "format": "uuid"
          },
          "uploadUrl": {
            "type": "string",
            "format": "uri"
          },
          "uploadUrlExpiresAt": {
            "type": "string",
            "format": "date-time"
          },
          "requiredHeaders": {
            "type": "object",
            "properties": {
              "Content-Type": {
                "type": "string",
                "enum": ["application/octet-stream"]
              },
              "Content-Length": {
                "type": "string",
                "pattern": "^\\d+$"
              }
            },
            "required": ["Content-Type", "Content-Length"],
            "additionalProperties": false
          }
        },
        "required": ["artifactId", "uploadUrl", "uploadUrlExpiresAt", "requiredHeaders"],
        "additionalProperties": false
      },
      "SdkTraceArtifactUploadRequest": {
        "type": "object",
        "properties": {
          "logicalName": {
            "type": "string",
            "minLength": 1,
            "maxLength": 160
          },
          "producerTool": {
            "type": "string",
            "minLength": 1,
            "maxLength": 128
          },
          "producerVersion": {
            "type": "string",
            "minLength": 1,
            "maxLength": 128
          },
          "byteLength": {
            "type": "integer",
            "exclusiveMinimum": 0,
            "maximum": 5363340410
          },
          "sha256": {
            "type": "string",
            "pattern": "^[a-f0-9]{64}$/i"
          }
        },
        "required": ["logicalName", "byteLength", "sha256"],
        "additionalProperties": false
      },
...
    },
...
  },
  "paths": {
...
    "/sdk/traces/{traceId}/artifacts": {
      "post": {
        "tags": ["Endpoints"],
        "operationId": "iniciar-carga-de-artefacto-de-traza",
        "summary": "Solicitar una carga directa de artefacto de validación",
        "description": "Crea una carga privada de un archivo fuente vinculado a una traza vigente. El servidor fija el tamaño, el SHA-256 y el nombre lógico declarados y devuelve una URL de R2 de una hora para un único PUT de `application/octet-stream`. El cliente no debe reenviar `Authorization` al almacenamiento ni seguir redirecciones a otro origen. Al completar, el servidor verifica bytes, hash y estructura protobuf nativa de Perfetto. Se aceptan hasta 4.995 GiB por PUT directo; cargas mayores requieren multipart y no están habilitadas.\n\n## Errores\n\n| Estado | `code` | `message` | Cuándo ocurre |\n| --- | --- | --- | --- |\n| `401` | `invalidCredential` | La credencial no es válida. | Falta el encabezado `Authorization`, el secreto no tiene el formato `ayni_sk_…`, no corresponde a ninguna credencial o la aplicación de la credencial está archivada. |\n| `401` | `credentialRevoked` | La credencial fue revocada. Genera una nueva credencial para continuar. | Un administrador revocó o regeneró la credencial. |\n| `400` | `invalidArtifact` | La solicitud del artefacto no es válida. | Se requiere un nombre lógico, tamaño y SHA-256 válidos. |\n| `400` | `invalidArtifactContent` | El archivo no coincide con el tamaño, SHA-256 o formato permitidos. | Solo se aceptan trazas nativas de Perfetto en formato protobuf binario. |\n| `403` | `telemetryDisabled` | La política de telemetría de esta aplicación está deshabilitada. | La política debe seguir habilitada al crear y completar la carga. |\n| `404` | `traceNotFound` | No encontramos esta traza o artefacto. | La traza debe existir, pertenecer a la aplicación de la credencial y no haber vencido. |\n| `409` | `artifactUploadExpired` | La carga del artefacto venció. | La URL firmada tiene una vigencia limitada. |\n| `409` | `artifactUploadInProgress` | La carga del artefacto aún se está verificando. | Solo una verificación por artefacto se ejecuta a la vez; vuelve a intentar cuando termine. |\n| `409` | `artifactUploadChanged` | El archivo cambió durante la verificación. | El objeto de staging se reemplazó durante la comprobación; solicita otra carga. |\n| `413` | `artifactTooLarge` | El artefacto supera el tamaño máximo permitido. | La carga directa de una sola operación admite hasta 5363340410 bytes (4.995 GiB). |\n| `413` | `invalidArtifact` | La solicitud del artefacto no es válida. | El cuerpo del endpoint de confirmación supera el límite permitido. |",
        "security": [
          {
            "sdkCredential": []
          }
        ],
        "parameters": [
          {
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "required": true,
            "name": "traceId",
            "in": "path"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SdkTraceArtifactUploadRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Carga directa autorizada; usa solo los encabezados devueltos en el PUT a R2.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SdkTraceArtifactUploadIntent"
                }
              }
            }
          },
          "400": {
            "description": "Solicitud no válida.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "description": "Mensaje en español para mostrar o registrar."
                    },
                    "code": {
                      "type": "string",
                      "enum": ["invalidArtifact", "invalidArtifactContent"],
                      "description": "Código estable para decidir qué hacer."
                    }
                  },
                  "required": ["message", "code"]
                },
                "examples": {
                  "invalidArtifact": {
                    "description": "Se requiere un nombre lógico, tamaño y SHA-256 válidos.",
                    "value": {
                      "message": "La solicitud del artefacto no es válida.",
                      "code": "invalidArtifact"
                    }
                  },
                  "invalidArtifactContent": {
                    "description": "Solo se aceptan trazas nativas de Perfetto en formato protobuf binario.",
                    "value": {
                      "message": "El archivo no coincide con el tamaño, SHA-256 o formato permitidos.",
                      "code": "invalidArtifactContent"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Credencial no válida o revocada.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "description": "Mensaje en español para mostrar o registrar."
                    },
                    "code": {
                      "type": "string",
                      "enum": ["invalidCredential", "credentialRevoked"],
                      "description": "Código estable para decidir qué hacer."
                    }
                  },
                  "required": ["message", "code"]
                },
                "examples": {
                  "invalidCredential": {
                    "description": "Falta el encabezado `Authorization`, el secreto no tiene el formato `ayni_sk_…`, no corresponde a ninguna credencial o la aplicación de la credencial está archivada.",
                    "value": {
                      "message": "La credencial no es válida.",
                      "code": "invalidCredential"
                    }
                  },
                  "credentialRevoked": {
                    "description": "Un administrador revocó o regeneró la credencial.",
                    "value": {
                      "message": "La credencial fue revocada. Genera una nueva credencial para continuar.",
                      "code": "credentialRevoked"
                    }
                  }
                }
              }
            }
          },
          "403": {
            "description": "La política de telemetría no está habilitada.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "description": "Mensaje en español para mostrar o registrar."
                    },
                    "code": {
                      "type": "string",
                      "enum": ["telemetryDisabled"],
                      "description": "Código estable para decidir qué hacer."
                    }
                  },
                  "required": ["message", "code"]
                },
                "examples": {
                  "telemetryDisabled": {
                    "description": "La política debe seguir habilitada al crear y completar la carga.",
                    "value": {
                      "message": "La política de telemetría de esta aplicación está deshabilitada.",
                      "code": "telemetryDisabled"
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Recurso no disponible.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "description": "Mensaje en español para mostrar o registrar."
                    },
                    "code": {
                      "type": "string",
                      "enum": ["traceNotFound"],
                      "description": "Código estable para decidir qué hacer."
                    }
                  },
                  "required": ["message", "code"]
                },
                "examples": {
                  "traceNotFound": {
                    "description": "La traza debe existir, pertenecer a la aplicación de la credencial y no haber vencido.",
                    "value": {
                      "message": "No encontramos esta traza o artefacto.",
                      "code": "traceNotFound"
                    }
                  }
                }
              }
            }
          },
          "409": {
            "description": "El identificador ya se usó con otros datos.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "description": "Mensaje en español para mostrar o registrar."
                    },
                    "code": {
                      "type": "string",
                      "enum": [
                        "artifactUploadExpired",
                        "artifactUploadInProgress",
                        "artifactUploadChanged"
                      ],
                      "description": "Código estable para decidir qué hacer."
                    }
                  },
                  "required": ["message", "code"]
                },
                "examples": {
                  "artifactUploadExpired": {
                    "description": "La URL firmada tiene una vigencia limitada.",
                    "value": {
                      "message": "La carga del artefacto venció.",
                      "code": "artifactUploadExpired"
                    }
                  },
                  "artifactUploadInProgress": {
                    "description": "Solo una verificación por artefacto se ejecuta a la vez; vuelve a intentar cuando termine.",
                    "value": {
                      "message": "La carga del artefacto aún se está verificando.",
                      "code": "artifactUploadInProgress"
                    }
                  },
                  "artifactUploadChanged": {
                    "description": "El objeto de staging se reemplazó durante la comprobación; solicita otra carga.",
                    "value": {
                      "message": "El archivo cambió durante la verificación.",
                      "code": "artifactUploadChanged"
                    }
                  }
                }
              }
            }
          },
          "413": {
            "description": "El cuerpo supera el tamaño máximo permitido.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "description": "Mensaje en español para mostrar o registrar."
                    },
                    "code": {
                      "type": "string",
                      "enum": ["artifactTooLarge", "invalidArtifact"],
                      "description": "Código estable para decidir qué hacer."
                    }
                  },
                  "required": ["message", "code"]
                },
                "examples": {
                  "artifactTooLarge": {
                    "description": "La carga directa de una sola operación admite hasta 5363340410 bytes (4.995 GiB).",
                    "value": {
                      "message": "El artefacto supera el tamaño máximo permitido.",
                      "code": "artifactTooLarge"
                    }
                  },
                  "invalidArtifact": {
                    "description": "El cuerpo del endpoint de confirmación supera el límite permitido.",
                    "value": {
                      "message": "La solicitud del artefacto no es válida.",
                      "code": "invalidArtifact"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
  },
...
}